Feature/OPC-UA Vendor Supports (#67)
* Renaming: TMS->DAQ

In opc-ua we are also calling the spec daq. Because of this we should also not use the name TMS there.

* Remove options for building daq mandatory opc-ua specs

For the openDAQ SDK it makes no sense to switch the options to false, because these types need to be known,
so that variables from a opc-ua server a visible on openDAQ client side.

* Add Vendor Spec support for openDAQ Clients

If an electronic supplier add some more definition to opc-ua based on the openDAQ specs, it
is possible to add these via vendor spec defintion. For this a vendor needs to add
the source of the opc-ua files, Types, nodeIds and the NodeSet2 itself.

* Add cmake generation of endpoint nodeset register methods

* Fix opc-ua types were not added to opc-ua server.

Tests are running again.
Remove specific registerCustomTypes call for HBK Spec.

* Improve PR based on Feedback

Create separate CMakeLists.txt for HBK vendor spec.
Add compile option for the vendor spec.

* Register only data types if the server supports the namespaces

This is important because if all namespaces are added which are generated by the SDK it could come to namespace index missmatch.
It is very importnat that only the namepsaces provided by the opc-ua server are added and in the right order.

---------

Co-authored-by: Jaka Mohorko <jaka.mohorko@dewesoft.com>